rename file to be consistent with component name
